move hocs to core and add unit test
