+++ packages/bodiless-core/__tests__/hoc.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
 import { observable } from 'mobx';
-import { withContextActivator, withLocalContextMenu, withNodeDataHandlers } from '../src/hoc';
+import { withContextActivator, withLocalContextMenu, withNodeDataHandlers, withFlowToggle } from '../src/hoc';
 
 const TestComponent = ({ element: Element }: any) => (
   <Element><div>Test Component</div></Element>
@@ -33,5 +33,21 @@ describe('withNodeDataHandlers', () => {
     const DataComponent = withNodeDataHandlers(data)(TestComponent);
     const withProps = shallow(<DataComponent />);
     expect(withProps.props().componentData).toEqual(values);
+  });
+});
+
+describe('withFlowToggle', () => {
+  it('renders the correct component based on the toggle function', () => {
+    const A = () => <div id="A" />;
+    const B = () => <div id="B" />;
+    const createToggleFunc = (isOn: boolean) => () => isOn;
+    const RenderA = withFlowToggle(createToggleFunc(true))(A, B);
+    const RenderB = withFlowToggle(createToggleFunc(false))(A, B);
+    const wrapper = shallow(<RenderA />);
+    expect(wrapper.find('#A')).toHaveLength(1);
+    expect(wrapper.find('#B')).toHaveLength(0);
+    const wrapperB = shallow(<RenderB />);
+    expect(wrapperB.find('#A')).toHaveLength(0);
+    expect(wrapperB.find('#B')).toHaveLength(1);
   });
 });
