+++ packages/bodiless-core/src/hoc.tsx
@@ -13,7 +13,7 @@
  */
 
 import { observer } from 'mobx-react-lite';
-import React, { ComponentType as CT } from 'react';
+import React, { ComponentType, ComponentType as CT } from 'react';
 import { flowRight, omit } from 'lodash';
 import { useContextActivator, useEditContext } from './hooks';
 import { useNodeDataHandlers } from './NodeProvider';
@@ -88,3 +88,27 @@ export const withPageContext = <P extends object>({
       </PageContextProvider>
     );
   };
+
+type ToggleHook = (props: any) => boolean;
+
+export const withFlowToggle = (useToggle: ToggleHook) => <
+  P extends object,
+  Q extends object
+>(
+    On: ComponentType<P>,
+    Off: ComponentType<Q>,
+  ) => observer((props: P & Q) => (useToggle(props) ? <On {...props} /> : <Off {...props} />));
+
+export const ifToggledOn = (useToggle: ToggleHook) => <H extends Function>(
+  ...hocs: Function[]
+) => (
+    Component: ComponentType<any>,
+  //  @ts-ignore Ex  ct ed at least 1  rg uments, but got 0 or more.ts(2557)
+  ) => withFlowToggle(useToggle)(flowRight(...hocs)(Component), Component);
+
+export const ifToggledOff = (useToggle: ToggleHook) => <H extends Function>(
+  ...hocs: Function[]
+) => (
+    Component: ComponentType<any>,
+  // @ts-ignore Ex  ct ed at least   a rguments, but got 0 or more.ts(2557)
+  ) => withFlowToggle(useToggle)(Component, flowRight(...hocs)(Component));
